Trocano a distância do tipo double para decimal

+++ Models/PetShopFeliz.cs
@@ -7,6 +7,6 @@
         public decimal PrecoBanhoG { get; set; }
         public decimal PrecoBanhoP { get; set; }
 
-        public double Distancia { get; set; }
+        public decimal Distancia { get; set; }
     }
 }
